chore(name_search): split highlight module into its own file
Matches the codebase convention (mod.rs/foo.rs alongside foo/test.rs) used
by chat, document, and project in this crate — the highlight helper was
still inline in lib.rs with an inline test module.
