+++ rust/cloud-storage/name_search/src/lib.rs
@@ -3,10 +3,12 @@
 
 mod chat;
 mod document;
+mod highlight;
 mod project;
 
 pub use chat::*;
 pub use document::*;
+pub use highlight::*;
 pub use models_opensearch::SearchEntityType;
 use models_search_cursor::{PaginatedResult, SearchCursorAttributes};
 pub use project::*;
@@ -24,66 +26,6 @@ pub fn escape_regex(term: &str) -> String {
         escaped.push(c);
     }
     escaped
-}
-
-/// Applies the same `<macro_em>` name-highlight replacement that the Postgres
-/// name-search queries apply via `regexp_replace(..., 'gi')`, but against an
-/// in-memory name string. Returns `None` when the term is empty or the name
-/// does not contain the term (case-insensitive).
-pub fn highlight_name(name: &str, term: &str) -> Option<String> {
-    let term = term.trim();
-    if term.is_empty() {
-        return None;
-    }
-    let re = regex::Regex::new(&format!("(?i)({})", escape_regex(term))).ok()?;
-    if !re.is_match(name) {
-        return None;
-    }
-    Some(re.replace_all(name, "<macro_em>$1</macro_em>").into_owned())
-}
-
-#[cfg(test)]
-mod highlight_test {
-    use super::highlight_name;
-
-    #[test]
-    fn returns_none_for_empty_term() {
-        assert!(highlight_name("testingfoop", "").is_none());
-        assert!(highlight_name("testingfoop", "   ").is_none());
-    }
-
-    #[test]
-    fn returns_none_when_name_does_not_match() {
-        assert!(highlight_name("unrelated", "test").is_none());
-    }
-
-    #[test]
-    fn wraps_substring_matches_case_insensitively() {
-        assert_eq!(
-            highlight_name("testingfoop", "test").as_deref(),
-            Some("<macro_em>test</macro_em>ingfoop")
-        );
-        assert_eq!(
-            highlight_name("MD CHECKBOX LIST TEST", "test").as_deref(),
-            Some("MD CHECKBOX LIST <macro_em>TEST</macro_em>")
-        );
-    }
-
-    #[test]
-    fn wraps_all_occurrences() {
-        assert_eq!(
-            highlight_name("test of a test", "test").as_deref(),
-            Some("<macro_em>test</macro_em> of a <macro_em>test</macro_em>")
-        );
-    }
-
-    #[test]
-    fn escapes_regex_specials_in_term() {
-        assert_eq!(
-            highlight_name("plan (v2) draft", "(v2)").as_deref(),
-            Some("plan <macro_em>(v2)</macro_em> draft")
-        );
-    }
 }
 
 /// Errors for name search crate
